chore: rename overrided.ts to overridden.ts
